Fixed jumping cursor due to implementing "Blur"-Logic
As soon as the Keyboard loses focus on the Editor (blur), the value will be set.

+ added new Method - setBasePath
